split menu scene into two, now goes from menu->intro->map selection. Pause button is implemented into Norse Minigame.
