+++ Positively Dead/Assets/Scenes/MapSelect.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: fbe3d878e3c594b4cb6b63fc8d07224e
-folderAsset: yes
+guid: 1729131d63bb50944bf8a0ae0c7eef61
 DefaultImporter:
   externalObjects: {}
   userData:
