+++ Positively Dead/ProjectSettings/EditorBuildSettings.asset
@@ -12,6 +12,9 @@ EditorBuildSettings:
     path: Assets/Scenes/Intro.unity
     guid: 08439e69955cb45be81991b1561e5643
   - enabled: 1
+    path: Assets/Scenes/MapSelect.unity
+    guid: 1729131d63bb50944bf8a0ae0c7eef61
+  - enabled: 1
     path: Assets/Scenes/Island 1 - NORSE.unity
     guid: b457f45748f51354b80424c20bc7a918
   - enabled: 1
